+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2019-06-08
+### Added
+- `Color::Secondary`.
+### Changed
+- Many other changes related to keyboards because of new button types. Please see [`keyboard` module documentation](https://docs.rs/vk-bot/2/vk_bot/keyboard/index.html) and [VK API bot keyboard documentation](https://vk.com/dev/bots_docs_3) for more information.
+### Removed
+- `Color::Default`.
+
 ## [1.0.0] - 2019-05-11
 ### Changed
 - Update to rvk **v0.12.0**.